+++ src/helper/types.ts
@@ -1,5 +1,37 @@
 import { PackageDir, NamedPackageDir } from '@salesforce/core';
-import { MetadataInfo } from 'jsforce';
+import { QueryResult } from 'jsforce';
+
+export declare type DateString = string & {
+  __DateBrand: never;
+};
+export declare type BlobString = string & {
+  __BlobBrand: never;
+};
+export declare type Address = {
+  city: string | null;
+  country: string | null;
+  geocodeAccuracy: string | null;
+  latitude: number | null;
+  longitude: number | null;
+  postalCode: string | null;
+  state: string | null;
+  street: string | null;
+};
+
+export declare type SObjectFieldType = number | boolean | DateString | BlobString | string | Address;
+
+export interface SObjectDefinition<N extends string = string> {
+  Name: N;
+  Fields: {
+      [name: string]: SObjectFieldType | null;
+  };
+  ParentReferences: {
+      [name: string]: SObjectDefinition | null;
+  };
+  ChildRelationships: {
+      [name: string]: SObjectDefinition;
+  };
+}
 
 export enum ComponentStatus {
   Created = 'Created',
@@ -73,7 +105,7 @@ export interface ExternalDataSourceMeta {
   type?: string;
   [x: string | number | symbol]: unknown;
 }
-export interface CustomMetadata extends MetadataInfo {
+export interface CustomMetadata  {
   Id?: string;
   DeveloperName?: string;
   MasterLabel?: string;
@@ -114,6 +146,7 @@ export interface ProjectJsonParsed {
 export interface PackageDirParsed extends PackageDir {
   name?: string;
   fullPath?: string;
+  ignoreOnStage?: string[];
 }
 export type FileResponse = FileResponseSuccess | FileResponseFailure;
 
@@ -201,7 +234,6 @@ export interface ApexClass {
 }
 
 export interface ApexTestQueueItem {
-  attributes: [Object];
   Id?: string;
   ApexClass?: ApexClass;
   ApexClassId?: string;
@@ -212,14 +244,12 @@ export interface ApexTestQueueItem {
 }
 
 export interface ApexCodeCoverageAggregate {
-  attributes: [Object];
   ApexClassOrTrigger?: ApexClassOrTrigger;
   NumLinesCovered?: number;
   NumLinesUncovered?: number;
 }
 
 export interface ApexTestResult {
-  attributes: [Object];
   ApexClass?: ApexClass;
   Outcome?: string;
   MethodName?: string;
@@ -274,3 +304,122 @@ export interface CodeCoverageWarnings {
   name?: string;
   namespace: {}
 }
+
+export interface SingleMergeRequest {
+  id: number;
+  packages: Map<string, GitPackageInfos>;
+  labels: string;
+  commitSHA: string;
+  author: string;
+  username: string;
+  title: string;
+  web_url: string;
+  createdat: string;
+  description: string;
+  merged_by: string;
+  reviewer: string;
+  isBuild: boolean;
+  isOnFT: boolean;
+  isOnSIT: boolean;
+  isOnPreProd: boolean;
+  isOnProd: boolean;
+  isOnEWISIT: boolean;
+  mergeStatus: string;
+  milestone: string;
+}
+
+export interface GitPackageInfos {
+  package: string;
+  version: string;
+  srcChange: boolean;
+  releaseTag: string;
+  commitId: string;
+  isOnFT: boolean;
+  isOnSIT: boolean;
+  isOnPreProd: boolean;
+  isOnProd: boolean;
+  isOnEWISIT: boolean;
+}
+
+export interface PackageChange{
+  old_path?: string;
+  new_path?: string;
+}
+
+export interface SfpowerscriptsArtifact2{
+  Id?: string;
+  Name?: string;
+  Version__c?: string;
+  CommitId__c?: string;
+}
+
+export interface JobDetails{
+  name: string;
+  alias: string;
+  jobId: number;
+  status: string;
+  webUrl: string;
+  packages: string[];
+}
+
+export interface UnassignPackage {
+  Package: string;
+  Id?: string;
+  UnassignKeys: UnassignKeys[];
+}
+
+export interface UnassignKeys {
+  Type: string;
+  Component: string;
+  ParentObject?: string
+  IsRemoved?: boolean;
+  Message?: string
+}
+
+export interface ProjectValidationOutput {
+  Process: string;
+  Package: string;
+  Message: string;
+}
+
+export type MetadataPackageVersions = {
+  Id: string;
+  MajorVersion: string;
+  MinorVersion: string;
+  PatchVersion: string;
+  BuildNumber: string;
+  SystemModstamp: Date;
+}
+
+export type MetadataPackage = {
+  SObjects: {
+    [name: string]: SObjectDefinition;
+  };
+  Name: string;
+  MetadataPackageVersions: QueryResult<MetadataPackageVersions>;
+}
+
+export type MetadataPackageVersion = {
+  id: string;
+  name: string;
+  version: string;
+  modifiedDate: Date;
+}
+
+export type Dependencies = {
+  ids?: SubscriberPackageVersionId[];
+};
+
+export type SubscriberPackageVersion = {
+  Dependencies: Dependencies;
+}
+
+export type SubscriberPackageVersionId = {
+  subscriberPackageVersionId: string;
+};
+
+export type BranchCreateResponse = {
+  commit: string;
+  version: string;
+};
+