+++ CHANGELOG.md
@@ -2,6 +2,219 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](https://github.com/eon-com/eon-sfdx/compare/v1.16.0...v2.0.0) (2024-03-15)
+
+
+### Features
+
+* **eon-com:** breaking change to oclif ([4aeaedc](https://github.com/eon-com/eon-sfdx/commit/4aeaedc810501383f540c2f39f2801414a0c5328))
+* **eon-com:** unit tests ([d4395d2](https://github.com/eon-com/eon-sfdx/commit/d4395d2733a99807c3a5175ba2b6db400980e21c))
+* **eon-com:** update commit command ([b2a657e](https://github.com/eon-com/eon-sfdx/commit/b2a657e81edfc93fa353914bc82964b01cd0f2f5))
+
+
+## [1.16.0](https://github.com/eon-com/eon-sfdx/compare/v1.13.2...v1.16.0) (2023-11-22)
+
+
+### Features
+
+* **chore:** update salesforce core to latest ([df334ec](https://github.com/eon-com/eon-sfdx/commit/df334ecd44c6f07864935c182e72f31653b3c1c2))
+
+
+### Bug Fixes
+
+* (update timeout) new timeout for testclass execution ([62afaf8](https://github.com/eon-com/eon-sfdx/commit/62afaf867111d56cbbddd2c11c94cf83ce046fd4))
+* (update-metadata) parameter for artifatcs path ([271296f](https://github.com/eon-com/eon-sfdx/commit/271296f9956ff34a3a3364bea5be0480a54792be))
+* (validation commands) change testclass execution ([8c330f5](https://github.com/eon-com/eon-sfdx/commit/8c330f572389bf262035dae15a7e4d3d55fe0404))
+
+## [1.15.0](https://github.com/eon-com/eon-sfdx/compare/v1.13.2...v1.15.0) (2023-11-22)
+
+
+### Features
+
+* **chore:** update salesforce core to latest ([df334ec](https://github.com/eon-com/eon-sfdx/commit/df334ecd44c6f07864935c182e72f31653b3c1c2))
+
+
+### Bug Fixes
+
+* (update-metadata) parameter for artifatcs path ([271296f](https://github.com/eon-com/eon-sfdx/commit/271296f9956ff34a3a3364bea5be0480a54792be))
+* (validation commands) change testclass execution ([8c330f5](https://github.com/eon-com/eon-sfdx/commit/8c330f572389bf262035dae15a7e4d3d55fe0404))
+
+## [1.14.0](https://github.com/eon-com/eon-sfdx/compare/v1.13.2...v1.14.0) (2023-05-22)
+
+
+### Features
+
+* **chore:** update salesforce core to latest ([df334ec](https://github.com/eon-com/eon-sfdx/commit/df334ecd44c6f07864935c182e72f31653b3c1c2))
+
+
+### Bug Fixes
+
+* (update-metadata) parameter for artifatcs path ([271296f](https://github.com/eon-com/eon-sfdx/commit/271296f9956ff34a3a3364bea5be0480a54792be))
+
+### [1.13.3](https://github.com/eon-com/eon-sfdx/compare/v1.13.2...v1.13.3) (2023-04-27)
+
+
+### Bug Fixes
+
+* (update-metadata) parameter for artifatcs path ([271296f](https://github.com/eon-com/eon-sfdx/commit/271296f9956ff34a3a3364bea5be0480a54792be))
+
+### [1.13.2](https://github.com/eon-com/eon-sfdx/compare/v1.13.0...v1.13.2) (2023-04-19)
+
+
+### Bug Fixes
+
+* (project-validate) find correct path for branches <> main ([8a43daa](https://github.com/eon-com/eon-sfdx/commit/8a43daa52b3a28d7fda410597fddb5cb31955ae0))
+
+### [1.13.1](https://github.com/eon-com/eon-sfdx/compare/v1.13.0...v1.13.1) (2023-04-18)
+
+
+### Bug Fixes
+
+* (project-validate) find correct path for branches <> main ([8a43daa](https://github.com/eon-com/eon-sfdx/commit/8a43daa52b3a28d7fda410597fddb5cb31955ae0))
+
+## [1.13.0](https://github.com/eon-com/eon-sfdx/compare/v1.11.2...v1.13.0) (2023-04-18)
+
+
+### Features
+
+* (project-validation) add change flag to command ([a86d8c0](https://github.com/eon-com/eon-sfdx/commit/a86d8c0d8381e6b88fff2d61b2e7f926156a37e6))
+
+## [1.12.0](https://github.com/eon-com/eon-sfdx/compare/v1.11.2...v1.12.0) (2023-04-18)
+### Features
+
+* (project-validation) add change flag to command ([a86d8c0](https://github.com/eon-com/eon-sfdx/commit/a86d8c0d8381e6b88fff2d61b2e7f926156a37e6))
+
+### [1.11.2](https://github.com/eon-com/eon-sfdx/compare/v1.11.0...v1.11.2) (2023-03-29)
+
+### [1.11.1](https://github.com/eon-com/eon-sfdx/compare/v1.11.0...v1.11.1) (2023-03-27)
+### Bug Fixes
+
+* **project-validate:** fix dev versions and git push ([4cd62b1](https://github.com/eon-com/eon-sfdx/commit/4cd62b13320d05a5a092cc8cdc10fd1bae09fa15))
+
+## [1.11.0](https://github.com/eon-com/eon-sfdx/compare/v1.9.4...v1.11.0) (2023-03-13)
+
+
+### Features
+
+* **project-validation:** new logic for dep version. get versions from dev hub ([9252b45](https://github.com/eon-com/eon-sfdx/commit/9252b4557f761bf070f42983a6ee90b7aa53b3f7))
+
+
+### Bug Fixes
+
+* **project-validate:** check also dep from dep versions ([1cfdc01](https://github.com/eon-com/eon-sfdx/commit/1cfdc01063f1ea727b36db691bee39dc9f42e8cf))
+* **project-validation:** resolve output for default key string to boolean ([54f69d9](https://github.com/eon-com/eon-sfdx/commit/54f69d94a1db055123c426a6e2924fed16dceff6))
+
+### [1.10.1](https://github.com/eon-com/eon-sfdx/compare/v1.10.0...v1.10.1) (2023-03-13)
+
+### Bug Fixes
+
+* **gitlab-create-branch:** fix(git branch create command) new prompts ([#30](https://github.com/eon-com/eon-sfdx/issues/30)) ([5f1a185](https://github.com/eon-com/eon-sfdx/commit/5f1a1858d7c947e252b15aa2bdbda7682c21d68f))
+* **project-validation:** fix(command package validate) no error when uncovered lines == 0 ([#30](https://github.com/eon-com/eon-sfdx/issues/30)) ([2fd19af5](https://github.com/eon-com/eon-sfdx/commit/2fd19af50c9ffcf28539fa7f1d100f2a49f50ae8))
+* **project-validation:** fix(command-project-validate) change order for validate stages ([#30](https://github.com/eon-com/eon-sfdx/issues/30)) ([a571a93](https://github.com/eon-com/eon-sfdx/commit/a571a934968be6b400a39fb00c01d99654f51013))
+
+## [1.10.0](https://github.com/eon-com/eon-sfdx/compare/v1.9.4...v1.10.0) (2023-03-01)
+
+
+### Features
+
+* **project-validation:** new logic for dep version. get versions from dev hub ([9252b45](https://github.com/eon-com/eon-sfdx/commit/9252b4557f761bf070f42983a6ee90b7aa53b3f7))
+
+
+### Bug Fixes
+
+* **project-validate:** check also dep from dep versions ([1cfdc01](https://github.com/eon-com/eon-sfdx/commit/1cfdc01063f1ea727b36db691bee39dc9f42e8cf))
+* **project-validation:** resolve output for default key string to boolean ([54f69d9](https://github.com/eon-com/eon-sfdx/commit/54f69d94a1db055123c426a6e2924fed16dceff6))
+
+### [1.9.4](https://github.com/eon-com/eon-sfdx/compare/v1.9.3...v1.9.4) (2023-01-13)
+
+
+### Bug Fixes
+
+* **devkit:** Fixing an issue where multiple permission set assignments causes the script to fail  ([#25](https://github.com/eon-com/eon-sfdx/issues/25)) ([0170099](https://github.com/eon-com/eon-sfdx/commit/0170099f16ae9fb9994fd38913b426e391590100))
+* **project-validation:** resolve output error for method checkPackageOrder ([#26](https://github.com/eon-com/eon-sfdx/issues/26)) ([9461eda](https://github.com/eon-com/eon-sfdx/commit/9461eda265bf3d1171797d26de1881d69fe5bd25))
+
+### [1.9.3](https://github.com/eon-com/eon-sfdx/compare/v1.9.2...v1.9.3) (2022-12-19)
+
+
+### Bug Fixes
+
+* **project-validation:** resolve wrong order for managed packages ([993b6ce](https://github.com/eon-com/eon-sfdx/commit/993b6cefbc69493659e89e77cc9659269fe9368d))
+
+### [1.9.2](https://github.com/eon-com/eon-sfdx/compare/v1.9.1...v1.9.2) (2022-12-09)
+
+
+### Bug Fixes
+
+* **project-validate:** resolve table output issue from pipeline test ([c746be9](https://github.com/eon-com/eon-sfdx/commit/c746be95e3a60a8de283ef0650d49983249f2855))
+
+### [1.9.1](https://github.com/eon-com/eon-sfdx/compare/v1.9.0...v1.9.1) (2022-12-08)
+
+## [1.9.0](https://github.com/eon-com/eon-sfdx/compare/v1.7.0...v1.9.0) (2022-12-01)
+
+
+### Features
+
+* adding new command to create branch from org snapshot ([2a86434](https://github.com/eon-com/eon-sfdx/commit/2a86434c7b24d594c9423886d225286531422f95))
+
+
+### Bug Fixes
+
+* new project validate command ([601996c](https://github.com/eon-com/eon-sfdx/commit/601996cd8660b247a21358246003c8657d3a0696))
+* **project-validate:** fixed issue when parsing version numbers from packages ([6375380](https://github.com/eon-com/eon-sfdx/commit/63753800ce1a32765f1e98a30e2e71d17d709c73))
+* **validate-package:** fix issue where post deployment script path is found in artifacts ([d643dfa](https://github.com/eon-com/eon-sfdx/commit/d643dfaa8a5e877971d036e39d3028900604594c))
+* **validate-source:** fix issue where script is not executed correctly when called from artifact ([2f179f6](https://github.com/eon-com/eon-sfdx/commit/2f179f680294844dc765ba5a3c5530594f6fc8cd))
+
+## [1.8.0](https://github.com/eon-com/eon-sfdx/compare/v1.7.0...v1.8.0) (2022-11-17)
+
+
+### Features
+
+* adding new command to create branch from org snapshot ([2a86434](https://github.com/eon-com/eon-sfdx/commit/2a86434c7b24d594c9423886d225286531422f95))
+
+
+### Bug Fixes
+
+* new project validate command ([601996c](https://github.com/eon-com/eon-sfdx/commit/601996cd8660b247a21358246003c8657d3a0696))
+
+## [1.7.0](https://github.com/eon-com/eon-sfdx/compare/v1.6.1...v1.7.0) (2022-10-17)
+
+
+### Features
+
+* (new command) create table for error handling ([ce5dc4b](https://github.com/eon-com/eon-sfdx/commit/ce5dc4b77d768dff4680a850651da5358c5ffafe))
+
+
+### Bug Fixes
+
+* (command) activate:bot ([dc57da9](https://github.com/eon-com/eon-sfdx/commit/dc57da903170c061f13506906f9f383b01e511df))
+* (command) activate:bot ([30b3426](https://github.com/eon-com/eon-sfdx/commit/30b3426028c67c4042161576399130612b73bef7))
+* (command) package:validate ([3bb9459](https://github.com/eon-com/eon-sfdx/commit/3bb945979bd79504eea5c43dafcb809e305055f2))
+
+### [1.6.1](https://github.com/eon-com/eon-sfdx/compare/v1.6.0...v1.6.1) (2022-10-10)
+
+
+### Bug Fixes
+
+* **validate:** resolve duplicate testclass names ([83df52c](https://github.com/eon-com/eon-sfdx/commit/83df52c1931658228257523dcb14f3d6277f2d61))
+
+## [1.6.0](https://github.com/eon-com/eon-sfdx/compare/v1.5.1...v1.6.0) (2022-10-04)
+
+
+### Features
+
+* new command for static project analysis ([6f7d3a7](https://github.com/eon-com/eon-sfdx/commit/6f7d3a75dd1186c26eff798bf379c429cbf96666))
+
+
+### Bug Fixes
+
+* **validate:source:** resolve error where single errors are not displayed ([3f67c61](https://github.com/eon-com/eon-sfdx/commit/3f67c61adf45dcb720633002d7955035cddaff9b))
+
+### [1.5.1](https://github.com/eon-com/eon-sfdx/compare/v1.5.0...v1.5.1) (2022-09-28)
+
+### [1.4.2](https://github.com/eon-com/eon-sfdx/compare/v1.4.1...v1.4.2) (2022-09-20)
+
+### [1.4.1](https://github.com/eon-com/eon-sfdx/compare/v1.4.0...v1.4.1) (2022-09-20)
+
 ## [1.3.0](https://github.com/eon-com/eon-sfdx/compare/v1.3.0-beta.0...v1.3.0) (2022-09-09)
 
 ## [1.2.0-beta.7](https://github.com/eon-com/eon-sfdx/compare/v1.2.0-beta.6...v1.2.0-beta.7) (2022-08-31)